+++ CHANGES.md
@@ -12,6 +12,7 @@
 * Set an environment variable if a token has a `=` sign (Yuliang Tan)
 * Redirect Output with subprocess if a token has a `>` sign (Yuliang Tan)
 * Run the command at the background if a token has a `&` sign (Yuliang Tan)
+* Auto find file if a token has a `*` sign (Yuliang Tan)
 
 2016-07-08, Version 0.1.0
 =========================